# openapi.json
{"openapi":"3.1.0","info":{"title":"ID Service","description":"OpenID Connect Identity Provider for asynq.ru","version":"1.0.0"},"paths":{"/.well-known/openid-configuration":{"get":{"tags":["OIDC Discovery"],"summary":"Openid Configuration","description":"OpenID Connect Discovery endpoint","operationId":"openid_configuration__well_known_openid_configuration_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiscoveryResponse"}}}}}}},"/.well-known/jwks.json":{"get":{"tags":["OIDC JWKS"],"summary":"Jwks","operationId":"jwks__well_known_jwks_json_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/authorize":{"get":{"tags":["OIDC Authorization"],"summary":"Authorize","operationId":"authorize_authorize_get","parameters":[{"name":"client_id","in":"query","required":true,"schema":{"type":"string","title":"Client Id"}},{"name":"response_type","in":"query","required":true,"schema":{"type":"string","title":"Response Type"}},{"name":"redirect_uri","in":"query","required":true,"schema":{"type":"string","title":"Redirect Uri"}},{"name":"scope","in":"query","required":true,"schema":{"type":"string","title":"Scope"}},{"name":"state","in":"query","required":true,"schema":{"type":"string","minLength":1,"title":"State"}},{"name":"nonce","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nonce"}},{"name":"code_challenge","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code Challenge"}},{"name":"code_challenge_method","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code Challenge Method"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/token":{"post":{"tags":["OIDC Token"],"summary":"Token","operationId":"token_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_token_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/userinfo":{"get":{"tags":["OIDC UserInfo"],"summary":"Userinfo","description":"OpenID Connect UserInfo endpoint","operationId":"userinfo_userinfo_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfoResponse"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["OIDC UserInfo"],"summary":"Userinfo","description":"OpenID Connect UserInfo endpoint","operationId":"userinfo_userinfo_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfoResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/logout":{"post":{"tags":["OIDC Logout"],"summary":"Logout Post","description":"JSON-логаут.\nДействия:\n  - Если есть IdP-сессия: отозвать, очистить cookie, дернуть back-channel.\n  - Валидация post_logout_redirect_uri идёт по клиенту из id_token_hint.\n  - Возвращаем { ok: true, redirect_to: <uri>|null }. 302 нет.","operationId":"logout_post_logout_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["OIDC Logout"],"summary":"Logout Get","description":"Упрощённый GET-вариант.\nНикаких HTML/iframes. Без очистки cookie и без ревокации.\nЕсли redirect валиден по клиенту из id_token_hint — 302 туда (добавим state).\nИначе — 204 No Content.","operationId":"logout_get_logout_get","parameters":[{"name":"id_token_hint","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id Token Hint"}},{"name":"post_logout_redirect_uri","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Post Logout Redirect Uri"}},{"name":"state","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/revoke":{"post":{"tags":["OIDC Token"],"summary":"Revoke","operationId":"revoke_revoke_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_revoke_revoke_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register new user","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login-password":{"post":{"tags":["Authentication"],"summary":"Login Password","description":"Логин по паролю с JSON-телом.\nПоведение:\n  - Успех без state: { ok:true, redirect_to:null } + установка SSO-cookie.\n  - Успех с валидными state+client_id (pending в Redis): выдаём code, удаляем pending,\n    возвращаем { ok:true, redirect_to:\"<redirect_uri>?code=...&state=...\" } + cookie.\nОшибки: через HTTPException, их перехватит глобальный хендлер.","operationId":"login_password_api_auth_login_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPasswordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/forgot-password":{"post":{"tags":["Authentication"],"summary":"Forgot Password","description":"Request password reset","operationId":"forgot_password_api_auth_forgot_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/verify-reset":{"post":{"tags":["Authentication"],"summary":"Verify Reset","description":"Verify password reset code","operationId":"verify_reset_api_auth_verify_reset_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyResetRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyResetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/set-new-password":{"post":{"tags":["Authentication"],"summary":"Set New Password","description":"Set new password after reset","operationId":"set_new_password_api_auth_set_new_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetNewPasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetNewPasswordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/verify-email":{"post":{"tags":["Authentication"],"summary":"Verify Email","description":"Подтверждение e-mail по OTP коду.\nПоведение:\n  - Если есть pending /authorize по state и redirect валиден — выпускаем code,\n    привязываем sid, удаляем pending, возвращаем JSON с redirect_to.\n  - Иначе обычный JSON без redirect_to.\n  - ВО ВСЕХ случаях ставим SSO-cookie на этот JSON-ответ.","operationId":"verify_email_api_auth_verify_email_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/csrf":{"get":{"tags":["Authentication"],"summary":"Get Csrf","operationId":"get_csrf_api_auth_csrf_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/account/profile":{"get":{"tags":["Account"],"summary":"Get Profile","description":"Get current user profile","operationId":"get_profile_api_account_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}}}},"/api/account/change-email/request":{"post":{"tags":["Account"],"summary":"Request Email Change","description":"Request email change - sends verification code to new email","operationId":"request_email_change_api_account_change_email_request_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEmailRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeEmailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/account/change-email/confirm":{"post":{"tags":["Account"],"summary":"Confirm Email Change","description":"Confirm email change with verification code","operationId":"confirm_email_change_api_account_change_email_confirm_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailChangeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmEmailChangeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/account/change-password":{"post":{"tags":["Account"],"summary":"Change Password","description":"Change password for authenticated user","operationId":"change_password_api_account_change_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/account/delete":{"delete":{"tags":["Account"],"summary":"Delete Account","description":"Delete user account (soft delete)","operationId":"delete_account_api_account_delete_delete","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteAccountRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteAccountResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health","description":"Health check endpoint","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/ready":{"get":{"tags":["Health"],"summary":"Readiness","description":"Readiness check with database connection","operationId":"readiness_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_revoke_revoke_post":{"properties":{"token":{"type":"string","title":"Token"},"token_type_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Token Type Hint"},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["token"],"title":"Body_revoke_revoke_post"},"Body_token_token_post":{"properties":{"grant_type":{"type":"string","title":"Grant Type"},"code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code"},"redirect_uri":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Redirect Uri"},"code_verifier":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code Verifier"},"refresh_token":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Refresh Token"},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["grant_type"],"title":"Body_token_token_post"},"ChangeEmailRequest":{"properties":{"new_email":{"type":"string","format":"email","title":"New Email","description":"New email address"},"password":{"type":"string","title":"Password","description":"Current password for verification"}},"type":"object","required":["new_email","password"],"title":"ChangeEmailRequest","description":"Change email request schema"},"ChangeEmailResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Verification code sent to new email address"}},"type":"object","title":"ChangeEmailResponse","description":"Change email response schema"},"ChangePasswordRequest":{"properties":{"old_password":{"type":"string","title":"Old Password","description":"Current password"},"new_password":{"type":"string","minLength":8,"title":"New Password","description":"New password"},"new_password_confirm":{"type":"string","title":"New Password Confirm","description":"New password confirmation"}},"type":"object","required":["old_password","new_password","new_password_confirm"],"title":"ChangePasswordRequest","description":"Change password request schema"},"ChangePasswordResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Password changed successfully"}},"type":"object","title":"ChangePasswordResponse","description":"Change password response schema"},"ConfirmEmailChangeRequest":{"properties":{"new_email":{"type":"string","format":"email","title":"New Email"},"code":{"type":"string","pattern":"^\\d{4}$","title":"Code","description":"4-digit verification code"}},"type":"object","required":["new_email","code"],"title":"ConfirmEmailChangeRequest"},"ConfirmEmailChangeResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Email address updated successfully"}},"type":"object","title":"ConfirmEmailChangeResponse","description":"Confirm email change response schema"},"DeleteAccountRequest":{"properties":{"current_password":{"type":"string","title":"Current Password"},"confirmation":{"type":"string","pattern":"^DELETE$","title":"Confirmation","description":"Type DELETE to confirm"}},"type":"object","required":["current_password","confirmation"],"title":"DeleteAccountRequest","description":"Delete account request schema"},"DeleteAccountResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Account has been deleted"}},"type":"object","title":"DeleteAccountResponse","description":"Delete account response schema"},"DiscoveryResponse":{"properties":{"issuer":{"type":"string","title":"Issuer"},"authorization_endpoint":{"type":"string","title":"Authorization Endpoint"},"token_endpoint":{"type":"string","title":"Token Endpoint"},"userinfo_endpoint":{"type":"string","title":"Userinfo Endpoint"},"jwks_uri":{"type":"string","title":"Jwks Uri"},"end_session_endpoint":{"type":"string","title":"End Session Endpoint"},"response_types_supported":{"items":{"type":"string"},"type":"array","title":"Response Types Supported"},"grant_types_supported":{"items":{"type":"string"},"type":"array","title":"Grant Types Supported"},"scopes_supported":{"items":{"type":"string"},"type":"array","title":"Scopes Supported"},"id_token_signing_alg_values_supported":{"items":{"type":"string"},"type":"array","title":"Id Token Signing Alg Values Supported"},"subject_types_supported":{"items":{"type":"string"},"type":"array","title":"Subject Types Supported"},"token_endpoint_auth_methods_supported":{"items":{"type":"string"},"type":"array","title":"Token Endpoint Auth Methods Supported"},"code_challenge_methods_supported":{"items":{"type":"string"},"type":"array","title":"Code Challenge Methods Supported"},"claims_supported":{"items":{"type":"string"},"type":"array","title":"Claims Supported"},"frontchannel_logout_supported":{"type":"boolean","title":"Frontchannel Logout Supported"},"backchannel_logout_supported":{"type":"boolean","title":"Backchannel Logout Supported"},"backchannel_logout_session_supported":{"type":"boolean","title":"Backchannel Logout Session Supported"}},"type":"object","required":["issuer","authorization_endpoint","token_endpoint","userinfo_endpoint","jwks_uri","end_session_endpoint","response_types_supported","grant_types_supported","scopes_supported","id_token_signing_alg_values_supported","subject_types_supported","token_endpoint_auth_methods_supported","code_challenge_methods_supported","claims_supported","frontchannel_logout_supported","backchannel_logout_supported","backchannel_logout_session_supported"],"title":"DiscoveryResponse"},"ForgotPasswordRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"Email address"}},"type":"object","required":["email"],"title":"ForgotPasswordRequest","description":"Запрос на отправку кода для сброса пароля."},"ForgotPasswordResponse":{"properties":{"message":{"type":"string","title":"Message","default":"If an account exists with this email, a reset code has been sent"}},"type":"object","title":"ForgotPasswordResponse","description":"Ответ на запрос сброса пароля.\nТекст одинаковый, чтобы не раскрывать наличие аккаунта."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginPasswordRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"Email address"},"password":{"type":"string","title":"Password","description":"Password"},"remember_me":{"type":"boolean","title":"Remember Me","description":"Extended session duration","default":false},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id","description":"OIDC client_id for pending /authorize"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State","description":"State from pending /authorize"}},"type":"object","required":["email","password"],"title":"LoginPasswordRequest","description":"Запрос логина по паролю.\nJSON-only. Формы не используются.\nЕсли переданы client_id+state и в Redis есть pending authreq — вернём redirect_to."},"LoginPasswordResponse":{"properties":{"ok":{"type":"boolean","title":"Ok","default":true},"message":{"type":"string","title":"Message","default":"Login successful"},"redirect_to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Redirect To"}},"type":"object","title":"LoginPasswordResponse","description":"Ответ на логин по паролю."},"LogoutRequest":{"properties":{"id_token_hint":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id Token Hint","description":"ID Token пользователя для определения клиента и валидации post_logout_redirect_uri"},"post_logout_redirect_uri":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Post Logout Redirect Uri","description":"Куда редиректить после логаута, если валиден для клиента"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State","description":"Опциональное состояние, вернётся как параметр в redirect_to"}},"type":"object","title":"LogoutRequest","description":"JSON-запрос на логаут по OIDC.\nПримечание: post_logout_redirect_uri валидируется по клиенту из id_token_hint."},"LogoutResponse":{"properties":{"ok":{"type":"boolean","title":"Ok","default":true},"redirect_to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Redirect To"}},"type":"object","title":"LogoutResponse","description":"JSON-ответ логаута.\nredirect_to: абсолютный URL или null. Фронт сам выполнит переход, 302 нет."},"ProfileResponse":{"properties":{"email":{"type":"string","title":"Email"},"username":{"type":"string","title":"Username"},"email_verified":{"type":"boolean","title":"Email Verified"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"last_login_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Login At"}},"type":"object","required":["email","username","email_verified","created_at"],"title":"ProfileResponse","description":"User profile response schema"},"RegisterRequest":{"properties":{"username":{"type":"string","maxLength":30,"minLength":3,"title":"Username","description":"Username"},"email":{"type":"string","format":"email","title":"Email","description":"Email address"},"password":{"type":"string","minLength":8,"title":"Password","description":"Password"},"password_confirm":{"type":"string","title":"Password Confirm","description":"Password confirmation"}},"type":"object","required":["username","email","password","password_confirm"],"title":"RegisterRequest","description":"Запрос на регистрацию."},"RegisterResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"message":{"type":"string","title":"Message","default":"Registration successful. Please check your email for verification code."}},"type":"object","required":["user_id"],"title":"RegisterResponse","description":"Ответ на регистрацию."},"SetNewPasswordRequest":{"properties":{"user_id":{"type":"string","title":"User Id","description":"User ID from reset verification"},"reset_token":{"type":"string","title":"Reset Token","description":"Reset token from verification"},"new_password":{"type":"string","minLength":8,"title":"New Password","description":"New password"},"new_password_confirm":{"type":"string","title":"New Password Confirm","description":"New password confirmation"}},"type":"object","required":["user_id","reset_token","new_password","new_password_confirm"],"title":"SetNewPasswordRequest","description":"Установка нового пароля после успешной верификации сброса."},"SetNewPasswordResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Password has been reset successfully"}},"type":"object","title":"SetNewPasswordResponse","description":"Ответ на успешную установку нового пароля."},"UserInfoResponse":{"properties":{"sub":{"type":"string","title":"Sub"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"email_verified":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Email Verified"},"preferred_username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preferred Username"},"updated_at":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["sub"],"title":"UserInfoResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyEmailRequest":{"properties":{"user_id":{"type":"string","title":"User Id","description":"User ID from registration"},"code":{"type":"string","pattern":"^\\d{4}$","title":"Code","description":"4-digit verification code"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"}},"type":"object","required":["user_id","code"],"title":"VerifyEmailRequest","description":"Подтверждение e-mail по OTP-коду.\nДополнительно поддерживает state для завершения pending /authorize."},"VerifyEmailResponse":{"properties":{"ok":{"type":"boolean","title":"Ok","default":true},"message":{"type":"string","title":"Message","default":"Email verified successfully"},"redirect_to":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Redirect To"}},"type":"object","title":"VerifyEmailResponse","description":"Ответ на верификацию e-mail."},"VerifyResetRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"Email address"},"code":{"type":"string","pattern":"^\\d{4}$","title":"Code","description":"4-digit reset code"}},"type":"object","required":["email","code"],"title":"VerifyResetRequest","description":"Подтверждение кода для сброса пароля."},"VerifyResetResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"reset_token":{"type":"string","title":"Reset Token"}},"type":"object","required":["user_id","reset_token"],"title":"VerifyResetResponse","description":"Ответ с reset_token для установки нового пароля."}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}